update: DTO 분리
